requirements.txt
langchain
langchain-ollama
sentence-transformers
faiss-cpu
pandas
numpy
